Make double->int casts explicit.
Fixes VS13 compiler warnings. Not sure why g++ isn't catching them.

+++ src/Camera.cpp
@@ -51,7 +51,7 @@ double Camera::getHeight() {
 }
 
 Size Camera::getSize() {
-    return Size(getWidth(), getHeight());
+    return Size((int)getWidth(), (int)getHeight());
 }
 
 Mat Camera::getNextFrame() {
